use `kill -SIGUSR1 <pid>` to inspect stats of zipper, close #219
